Consolidate name utils into new `name-utils` file
