--- web-common/src/features/entity-management/getNextEntityId.ts
@@ -1,11 +1,0 @@
-export function getNextEntityName(
-  entityNames: Array<string>,
-  entityName: string
-): string {
-  const idx = entityNames.indexOf(entityName);
-  if (idx <= 0) {
-    return entityNames[idx + 1];
-  } else {
-    return entityNames[idx - 1];
-  }
-}
