+++ web-common/src/features/models/navigation/ModelMenuItems.svelte
@@ -33,7 +33,7 @@
   import { useQueryClient } from "@sveltestack/svelte-query";
   import { createEventDispatcher } from "svelte";
   import { deleteFileArtifact } from "../../entity-management/actions";
-  import { getName } from "../../entity-management/incrementName";
+  import { getName } from "../../entity-management/name-utils";
   import { useModelNames } from "../selectors";
 
   export let modelName: string;
