+++ web-local/src/lib/components/workspace/metrics-def/MetricsDefWorkspaceHeader.svelte
@@ -3,9 +3,9 @@
   import MetricsIcon from "@rilldata/web-common/components/icons/Metrics.svelte";
   import { notifications } from "@rilldata/web-common/components/notifications";
   import { renameFileArtifact } from "@rilldata/web-common/features/entity-management/actions";
+  import { isDuplicateName } from "@rilldata/web-common/features/entity-management/name-utils";
   import { useAllNames } from "@rilldata/web-common/features/entity-management/selectors";
   import { EntityType } from "@rilldata/web-common/features/entity-management/types";
-  import { isDuplicateName } from "@rilldata/web-common/features/entity-management/utils";
   import { useRuntimeServiceRenameFileAndReconcile } from "@rilldata/web-common/runtime-client";
   import {
     appQueryStatusStore,
